requirements.txt
llama-index
llama-index-core
llama-index-embeddings-openai
llama-index-llms-openai
llama-parse
flask
werkzeug
python-dotenv